% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allreg.R
\name{allreg}
\alias{allreg}
\title{Regressions For Causal Mediation Analysis Approaches}

\usage{

allreg(model, data, yreg, mreg, ereg, postcreg,
wmnomreg, wmdenomreg, outcome, event, exposure,
mediator, EMint = FALSE, prec, postc)

}

\arguments{

\item{model}{the causal mediation analysis approach. \code{rb}, \code{wb}, \code{iorw}, \code{msm}, \code{gformula}, \code{ne} are implemented:
\itemize{
\item{"\code{rb}": }{the regression-based approach by Valeri et al. (2013) and VanderWeele et al. (2014)}
\item{"\code{wb}": }{the weighting-based approach by VanderWeele et al. (2014)}
\item{"\code{iorw}": }{the inverse odd-ratio weighting approach by Tchetgen Tchetgen et al. (2013)}
\item{"\code{msm}": }{the marginal structural models by VanderWeele et al. (2009)}
\item{"\code{gformula}": }{the g-formula approach by Lin et al. (2017)}
\item{"\code{ne}": }{the natural effect model implemented by \link[medflex]{neModel}, Steen et al. (2017)}
}
}

\item{data}{data for regression fitting}

\item{outcome}{the variable name of the outcome}

\item{event}{applicable and required when \code{yreg} is \code{coxph}, \code{aft_exp}, or \code{aft_weibull}. The variable name of the event}

\item{exposure}{the variable name of the exposure}

\item{mediator}{a character or vector of variable name(s) of the mediator(s)}

\item{EMint}{a logical value applicable and required when \code{yreg} is character. If \code{TRUE}, the outcome regression formula includes the interaction(s) between the exposure and each of the mediator(s). Default is \code{FALSE}. }

\item{prec}{a character or vector of variable name(s) of the pre-exposure confounder(s)}

\item{postc}{a character or vector of variable name(s) of the post-exposure confounder(s). If not \code{NULL}, \code{model} can only be \code{msm} or \code{gformula}}

\item{yreg}{outcome regression. Can be a user-defined regression object or a character specifying the regression name. See \code{details} for more information about supported values.}

\item{mreg}{applicable and required when \code{model} is \code{rb}, \code{msm} or \code{gformula}. A list of mediator regression(s) with a length of \code{length(mediator)} and following the same mediator order in \code{mediator}. Each element can be a user-defined regression object or a character specifying the regression name. See \code{details} for more information about supported values.}

\item{ereg}{applicable when \code{model} is \code{wb}, \code{iorw} or \code{msm}. Exposure regression for calculating weights which is set to be \code{NULL} when exposure weighting is not applied and set to be a user-defined regression object or a character specifying the regression name when exposure weighting is applied. See \code{details} for more information about supported values.}

\item{postreg}{applicable and required when \code{model} is \code{gformula} and \code{postc} is not empty. A list of post-exposure confounder regression(s) with a length of \code{length(postc)} and following the same mediator order in \code{postc}. Can be a user-defined regression object or a character specifying the regression name. See \code{details} for more information about supported values.}

\item{wmnomreg}{applicable when \code{model} is \code{msm}. A list of mediator regressions for calculating nominators of weights for each mediator with a length of \code{length(mediator)} and following the same mediator order in \code{mediator}. Each element can be \code{NULL}, a user-defined regression object or a character specifying the regression name. See \code{details} for more information about supported values.}

\item{wmdenomreg}{applicable when \code{model} is \code{msm}. A list of mediator regressions for calculating denominators of weights for each mediator with a length of \code{length(mediator)} and following the same mediator order in \code{mediator}. Each element can be \code{NULL}, a user-defined regression object or a character specifying the regression name. See \code{details} for more information about supported values.}

}

\value{

\code{allreg} returns a list of regressions required by the selected causal mediation analysis approach. For \code{rb} and \code{msm}, a list of \code{outcome_regression} and \code{mediator_regression} is returned; for \code{wb}, a list of \code{outcome_regression} and \code{exposure_regression} is returned; for \code{iorw}, a list of \code{tot_regression} and \code{dir_regression} is returned; for \code{gformula}, a list of \code{outcome_regression}, \code{mediator_regression} and \code{postc_regression} is returned.

\item{outcome_regression}{the fitted outcome regression object,}
\item{mediator_regression}{a list of fitted mediator regression object(s),}
\item{exposure_regression}{the fitted exposure regression object,}
\item{postc_regression}{a list of fitted post-exposure confounder regression object(s),}
\item{tot_regression}{the fitted outcome regression object for the total effect,}
\item{dir_regression}{the fitted outcome regression object for the direct effect,}

...

}

\description{
Run all regressions required by 5 causal mediation analysis approaches including the regression-based approach by Valeri et al. (2013) and VanderWeele et al. (2014), the weighting-based approach by VanderWeele et al. (2014), the inverse odd-ratio weighting approach by Tchetgen Tchetgen et al. (2013), the marginal structural models by VanderWeele et al. (2009), the g-formula approach by Lin et al. (2017) and the natural effect model implemented by \link[medflex]{neModel}, Steen et al. (2017).
}

\details{

\code{allreg} supports the following user-defined regression objects and characters specifying the regression name:
\itemize{
\item{"\code{linear}": }{simple linear regression fitted by \link{lm}}
\item{"\code{logistic}": }{logistic regression fitted by \link{glm} with \code{family=binomial()}}
\item{"\code{loglinear}": }{log-linear regression fitted by \link{glm} with \code{family=binomial("log")}}
\item{"\code{poisson}": }{poisson regression fitted by \link{glm} with \code{family=poisson()}}
\item{"\code{quasipoisson}": }{quasi-poisson regression fitted by \link{glm} with \code{family=quasipoisson()}}
\item{"\code{negbin}": }{negative binomial regression fitted by \link{glm.nb}}
\item{"\code{multinomial}": }{multinomial regression fitted by \link[nnet]{multinom}}
\item{"\code{ordinal}": }{proportional odds logistic regression fitted by \link[MASS]{polr}}
\item{"\code{coxph}": }{proportional hazards regression fitted by \link[survival]{coxph}}
\item{"\code{aft_exp}": }{parametric survival regression fitted by \link[survival]{survreg} with \code{dist="exponential"}}
\item{"\code{aft_weibull}": }{parametric survival regression fitted by \link[survival]{survreg} with \code{dist="weibull"}}
\item{\code{user-defined regression objects}: }{regression objects fitted by \link{lm}, \link{glm}, \link{glm.nb}, \link[mgcv]{gam}, \link[nnet]{multinom}, \link[MASS]{polr}, \link[survival]{survreg}, \link[survival]{coxph}}
}

All values above are supported for \code{yreg}. For \code{mreg} and \code{postcreg}, \code{coxph}, \code{aft_exp}, \code{aft_weibull}, regression objects fitted by \code{survreg()} and \code{coxph()} are not supported.

Let \code{Y} denote \code{otucome}, \code{A} denote \code{exposure}, \code{M} denote \code{mediator}, \code{C} denote \code{prec}, \code{L} denote \code{postc}.

\code{model="wb"} applies only when the exposure is categorical; when \code{prec} is empty, \code{ereg} is \code{NULL}, otherwise a non-null \code{ereg} is required for calculating weights and the weights are calculated by \code{P(A=a)/P(A=a|C=C_i)}. Thus, \code{ereg} for \code{model="wb"} can be \code{NULL}, \code{logistic}, \code{loglinear}, \code{multinomial}, \code{ordinal} or a user-defined regression object for categorical responses.

\code{model="iorw"} applies only when the exposure is categorical and \code{ereg} is required for calculating weights, in which the weights are calculated by \code{P(A=0|M=M_i,C=C_i)/P(A=A_i|M=M_i,C=C_i)}. Thus, \code{ereg} for \code{model="iorw"} can be \code{logistic}, \code{loglinear}, \code{multinomial}, \code{ordinal} or a user-defined regression object for categorical responses.

For \code{model="msm"}, when \code{prec} is empty, \code{ereg} is \code{NULL}, otherwise \code{ereg} is required for calculating weights in which the weight for subject \code{i} for \code{A} is calculated by \code{P(A=A_i)/P(A=A_i|C=C_i)} and only categorical exposure is supported; for \code{p=1,...,length(mediator)}, weight for subject \code{i} for \code{mediator[p]} is calculated by \code{P(M_p=M_{p,i}|A=A_i)/P(M_p=M_{p,i}|A=A_i,
C=C_i,L=L_i,M_1=M_{1,i},...,M_{p-1}=M_{p-1,i})}; when the nominator is the same as the denominator in the weight for subject \code{i} for \code{mediator[p]}, \code{wmnomreg[p]} and \code{wmdenomreg[p]} can be \code{NULL}, otherwise non-null \code{wmnomreg[p]} and \code{wmdenomreg[p]} are required and only categorical \code{mediator[p]} is supported. Thus, \code{ereg}, \code{wmnomreg[p]} and \code{wmdenomreg[p]} for \code{model="msm"} can be \code{NULL}, \code{logistic}, \code{loglinear}, \code{multinomial}, \code{ordinal} or a user-defined regression object for categorical responses.
}

\examples{





}

\references{

Valeri L, Vanderweele TJ (2013). Mediation analysis allowing for exposure-mediator interactions and causal interpretation: theoretical assumptions and implementation with SAS and SPSS macros. Psychological Methods. 18(2): 137 - 150.

VanderWeele TJ, Vansteelandt S (2014). Mediation analysis with multiple mediators. Epidemiologic Methods. 2(1): 95 - 115.

Tchetgen Tchetgen EJ (2013). Inverse odds ratio-weighted estimation for causal mediation analysis. Statistics in medicine. 32: 4567 - 4580.

Nguyen QC, Osypuk TL, Schmidt NM, Glymour MM, Tchetgen Tchetgen EJ. Practical guidance for conducting mediation analysis with multiple mediators using inverse odds ratio weighting (2015). American Journal of Epidemiology. 181(5): 349 - 356.

VanderWeele TJ. Marginal structural models for the estimation of direct and indirect effects (2009). Epidemiology. 20(1): 18 - 26.

VanderWeele TJ, Tchetgen Tchetgen EJ (2017). Mediation analysis with time varying exposures and mediators. Journal of the Royal Statistical Society: Series B (Statistical Methodology). 79(3): 917 - 938.

Lin SH, Young J, Logan R, Tchetgen Tchetgen EJ, VanderWeele TJ (2017). Parametric mediational g-formula approach to mediation analysis with time-varying exposures, mediators, and confounders. Epidemiology. 28: 266 - 274.

Steen J, Loeys T, Moerkerke B, Vansteelandt S (2017). Medflex: an R package for flexible mediation analysis using natural effect models. Journal of Statistical Software. 76(11).
}



\author{

}
