% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cma2020}
\alias{cma2020}
\title{The Dataset cma2020}
\format{
A data frame with 1000 rows and 8 variables:
\describe{
  \item{ID}{the id number for each subject}
  \item{C1}{a continuous pre-exposure confounder}
  \item{C2}{a binary pre-exposure confounder}
  \item{A}{the binary exposure}
  \item{M1}{a binary mediator}
  \item{M2}{a categorical mediator with 3 levels}
  \item{contY}{a continuous outcome}
  \item{binY}{a binary outcome}
}
}
\usage{
cma2020
}
\description{
An example dataset in the CMAverse package
}
\keyword{datasets}
