% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmdag.R
\name{cmdag}
\alias{cmdag}
\title{DAG For Causal Mediation Analysis}

\usage{
cmdag(outcome, exposure, mediator, prec, postc,
x.outcome = 4, x.exposure = 0, x.mediator = 2,
x.prec = 2, x.postc = 2, y.outcome = 0, y.exposure = 0,
y.mediator = 1, y.prec = 2, y.postc = -0.5,
...)
}

\arguments{
\item{outcome}{the variable name of the outcome}

\item{exposure}{the variable name of the exposure}

\item{mediator}{a character or vector of variable name(s) of the mediator(s)}

\item{prec}{a character or vector of variable name(s) of the pre-exposure confounder(s)}

\item{postc}{a character or vector of variable name(s) of the post-exposure confounder(s)}

\item{x.outcome}{x coordinate of the outcome. Default is \code{4}.}

\item{x.exposure}{x coordinate of the exposure. Default is \code{0}.}

\item{x.mediator}{x coordinate of the mediator. Default is \code{2}.}

\item{x.prec}{x coordinate of the pre-exposure confounder. Default is \code{2}.}

\item{x.postc}{x coordinate of the post-exposure confounder. Default is \code{2}.}

\item{y.outcome}{y coordinate of the outcome. Default is \code{0}.}

\item{y.exposure}{y coordinate of the exposure. Default is \code{0}.}

\item{y.mediator}{y coordinate of the mediator. Default is \code{1}.}

\item{y.prec}{y coordinate of the pre-exposure confounder. Default is \code{2}.}

\item{y.postc}{y coordinate of the post-exposure confounder. Default is \code{-0.5}.}

\item{...}{additional arguments passed to \code{ggdag()}. See \link[ggdag]{ggdag} for details.}
}

\description{
Create the directed acyclic graph(DAG) for causal mediation analysis
}
\details{
\code{outcome}, \code{exposure}, \code{mediator} are required and \code{prec}, \code{postc} are optional for creating the DAG.

For details about setting the coordinates, see \link[ggdag]{dagify}.
}

\examples{

## prec and postc are empty
cmdag(outcome = "Y", exposure = "A", mediator = c("M1", "M2"),
      prec = NULL, postc = NULL, node = TRUE, text_col = "white")

## postc is empty
cmdag(outcome = "Y", exposure = "A", mediator = c("M1", "M2"),
      prec = c("C1", "c2", "c3"), postc = NULL, node = FALSE, text_col = "black")

## prec and postc aren't empty
cmdag(outcome = "Y", exposure = "A", mediator = c("M1", "M2"),
      prec = c("C1", "c2", "c3"), postc = c("L1", "L2"), node = FALSE, text_col = "black")

}

\author{

}

