% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmest_multistate.R
\name{cmest_multistate}
\alias{cmest_multistate}
\title{Causal Mediation Analysis with the Multistate Approach}
\usage{
cmest_multistate(
  data = NULL,
  outcome = NULL,
  yevent = NULL,
  mediator = NULL,
  mevent = NULL,
  exposure = NULL,
  EMint = NULL,
  basec = NULL,
  basecval = NULL,
  ymreg = "coxph",
  astar = NULL,
  a = NULL,
  nboot = 200,
  bh_method = "breslow",
  s = NULL,
  multistate_seed = 123,
  n_workers = NULL
)
}
\arguments{
\item{data}{a data frame (or object coercible by \link{as.data.frame} to a data frame) 
containing the variables in the model.}

\item{outcome}{variable name of the outcome.}

\item{yevent}{variable name of the event for the outcome.}

\item{mediator}{variable name of the mediator.}

\item{mevent}{Event indicator for the mediator in multistate modeling.}

\item{exposure}{variable name of the exposure.}

\item{EMint}{a logical value. \code{TRUE} indicates there is 
exposure-mediator interaction in \code{yreg}.}

\item{basec}{a vector of variable names of the confounders.}

\item{basecval}{(required when \code{estimation} is \code{paramfunc} and \code{EMint} is \code{TRUE}) 
a list of values at which each confounder is conditioned on, following the order in \code{basec}. 
If \code{NULL}, the mean of each confounder is used.}

\item{ymreg}{type of multistate survival model to be used. Currently supporting coxph only.}

\item{astar}{the control value of the exposure.}

\item{a}{the treatment value of the exposure.}

\item{nboot}{(used when \code{inference} is \code{bootstrap}) the number of bootstraps applied. 
Default is 200.}

\item{bh_method}{Method for estimating baseline hazards in multistate modeling. Currently supporting "breslow" only.}

\item{s}{The time point(s) beyond which survival probability is interested in multistate modeling.}

\item{multistate_seed}{The seed to be used when generating bootstrap datasets for multistate modeling.}
}
\value{
The output is a list that consists of 4 elements:
\itemize{
\item{the model summary of the joint multistate Cox proportional hazards model fitted on the
original dataset}
\item{the point estimates of RD and SD for each of the
user-specified time points of interest on the original dataset}
\item{the summary of the bootstrapped RD, SD, and TE estimates for each of the
user-specified time point of interest, including the 2.5, 50, and
97.5th percentiles}
\item{the estimated RD, SD, TD for each of the
user-specified time point of interest for each bootstrap dataset}}
}
\description{
\code{cmest_multistate} is used to implement \emph{the multistate approach} by Valeri et al. (2023) 
for causal mediation analysis.
}
\details{
\strong{Assumptions of the multistate method}
\itemize{
\item{\emph{Consistency of potential outcomes:} }{For each i and each t, the survival in a world where we intervene, i.e., setting the time to treatment to a specific value t (via a fixed or stochastic intervention) is the same as the survival in the real world where we observe a time to treatment equal to t.}
\item{\emph{There is no unmeasured mediator-outcome confounding:} }{Given \code{exposure} and 
\code{basec}, \code{mediator} is independent of \code{outcome}.}
\item{\emph{Non-informative censoring of event times:} }{The observed censoring time is conditionally independent of all potential event times.}
\item{\emph{Positivity:} }{Each exposure-covariate combination has a non-zero probability of occurring.}
}
}
\examples{
\dontrun{
library(CMAverse)
multistate_out = cmest_multistate(data = sc_data, 
s = s_vec,
multistate_seed = 1,
exposure = 'A', mediator = 'M', outcome = 'S',
yevent = "ind_S", mevent = "ind_M",
basec = c("C1", "C2"),
basecval = c("C1" = "1", "C2" = as.character(mean(sc_data$C2))),
astar="0", a="1", 
nboot=1, EMint=F, 
bh_method = "breslow")
multistate_out
}

}
\references{
Valeri L, Proust-Lima C, Fan W, Chen JT, Jacqmin-Gadda H. A multistate approach for the study of interventions on an intermediate time-to-event in health disparities research. Statistical Methods in Medical Research. 2023;32(8):1445-1460.
}
