% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmest_rb.R
\name{cmest_rb}
\alias{cmest_rb}
\title{Causal Mediation Analysis with the Regression-Based Approach}
\usage{
cmest_rb(
  data = NULL,
  outcome = NULL,
  event = NULL,
  exposure = NULL,
  mediator = NULL,
  EMint = NULL,
  basec = NULL,
  yreg = NULL,
  mreg = NULL,
  estimation = "imputation",
  inference = "bootstrap",
  astar = NULL,
  a = NULL,
  mval = NULL,
  basecval = NULL,
  yval = NULL,
  nboot = 200,
  boot.ci.type = "per",
  casecontrol = FALSE,
  yrare = NULL,
  yprevalence = NULL,
  multimp = FALSE,
  args_mice = NULL
)
}
\arguments{
\item{data}{a data frame (or object coercible by \link{as.data.frame} to a data frame) 
containing the variables in the model.}

\item{outcome}{variable name of the outcome.}

\item{event}{(required when \code{yreg} is \code{coxph}, \code{aft_exp},
or \code{aft_weibull}) variable name of the event.}

\item{exposure}{variable name of the exposure.}

\item{mediator}{a vector of variable name(s) of mediator(s).}

\item{EMint}{a logical value. \code{TRUE} indicates there is 
exposure-mediator interaction in \code{yreg}.}

\item{basec}{a vector of variable names of confounders. See \code{Details}.}

\item{yreg}{outcome regression model. See \code{Details}.}

\item{mreg}{a list of mediator regression models following the order in \code{mediator}. See \code{Details}.}

\item{estimation}{estimation method. \code{paramfunc} and 
\code{imputation} are implemented (the first 4 letters are sufficient). Default is \code{imputation}. 
See \code{Details}.}

\item{inference}{inference method. \code{delta} and 
\code{bootstrap} are implemented (the first 4 letters are sufficient). Default is \code{bootstrap}. 
See \code{Details}.}

\item{astar}{the control value of the exposure.}

\item{a}{the treatment value of the exposure.}

\item{mval}{a list of values at which each mediator is controlled to calculate the \code{cde}, following the order in \code{mediator}.}

\item{basecval}{(required when \code{estimation} is \code{paramfunc} and \code{EMint} is \code{TRUE}) 
a list of values at which each confounder is conditioned on, following the order in \code{basec}. 
If \code{NULL}, the mean of each confounder is used.}

\item{yval}{(required when the outcome is categorical) the level of the outcome at which causal effects on the 
risk ratio scale are estimated. If \code{NULL}, the last level is used.}

\item{nboot}{(used when \code{inference} is \code{bootstrap}) the number of bootstraps applied. 
Default is 200.}

\item{boot.ci.type}{(used when \code{inference} is \code{bootstrap}) the type of bootstrap confidence interval. If \code{per}, percentile bootstrap
confidence intervals are estimated; if \code{bca}, bias-corrected and accelerated (BCa) bootstrap 
confidence intervals are estimated. Default is \code{per}.}

\item{casecontrol}{a logical value. \code{TRUE} indicates a case control study in which the
first level of the outcome is treated as the control and the second level of the outcome is 
treated as the case. Default is \code{FALSE}.}

\item{yrare}{(used when \code{casecontrol} is \code{TRUE}) a logical value. \code{TRUE} 
indicates the case is rare.}

\item{yprevalence}{(used when \code{casecontrol} is \code{TRUE}) the prevalence of the case.}

\item{multimp}{a logical value. If 
\code{TRUE}, conduct multiple imputations using the \link[mice]{mice} function. Default is 
\code{FALSE}.}

\item{args_mice}{a list of additional arguments passed to the \link[mice]{mice} function. See \link[mice]{mice}
for details.}

\item{x}{an object of class \code{cmest}.}

\item{object}{an object of class \code{cmest}.}

\item{digits}{minimal number of significant digits. See \link{print.default}.}
}
\value{
An object of classes \code{cmest} and \code{cmest_rb} is returned:
\item{call}{the function call,}
\item{data}{the data frame,}
\item{methods}{a list of methods used which may include \code{estimation}, \code{inference}, 
\code{nboot}, \code{boot.ci.type}, \code{casecontrol}, \code{yrare}, and \code{yprevalence},}
\item{variables}{a list of variables used which may include \code{outcome}, \code{event}, 
\code{exposure}, \code{mediator}, \code{EMint}, and \code{basec},}
\item{ref}{reference values used which may include \code{astar}, \code{a}, \code{mval}, 
\code{basecval} and \code{yval},}
\item{reg.input}{a list of regressions input,}
\item{reg.output}{a list of regressions output. If \code{multimp} is \code{TRUE}, 
reg.output contains regression models fitted by each imputed dataset,}
\item{multimp}{a list of arguments used for multiple imputation,}
\item{effect.pe}{point estimates of causal effects,}
\item{effect.se}{standard errors of causal effects,}
\item{effect.ci.low}{lower limits of the 95\% confidence intervals of causal effects,}
\item{effect.ci.high}{higher limits of the 95\% confidence intervals of causal effects,}
\item{effect.pval}{p-values of causal effects,}
...
}
\description{
\code{cmest_rb} is used to implement the \emph{the regression-based approach} 
by Valeri & VanderWeele (2013) and VanderWeele & Vansteelandt (2014) for causal mediation analysis 
with a single exposure, a single outcome, and a single or multiple mediators.
}
\details{
\strong{Assumptions of the regression-based approach}
\itemize{
\item{\emph{There is no unmeasured exposure-outcome confounding:} }{given \code{basec} and 
\code{postc}, \code{exposure} is independent of \code{outcome}.}
\item{\emph{There is no unmeasured mediator-outcome confounding:} }{given \code{exposure} and 
\code{basec}, \code{mediator} is independent of \code{outcome}.}
\item{\emph{There is no unmeasured exposure-mediator confounding:} }{given \code{basec}, 
\code{exposure} is independent of \code{mediator}.}
\item{\emph{There is no mediator-outcome confounder affected by the exposure:} }{there is no 
variable in \code{basec} affected by \code{exposure}.}
}


\strong{Regression models}

Each regression model in \code{yreg} and \code{mreg} can be specified by a fitted regression 
object or the character name of a regression model. 

\emph{The Character Name of a Regression Model:}
\itemize{
\item{\code{linear}: }{linear regression fitted by \link{glm} with \code{family = gaussian()}}
\item{\code{logistic}: }{logistic regression fitted by \link{glm} with \code{family = logit()}}
\item{\code{loglinear}: }{loglinear regression fitted by \link{glm} with 
\code{family = poisson()}}
\item{\code{poisson}: }{poisson regression fitted by \link{glm} with 
\code{family = poisson()}}
\item{\code{quasipoisson}: }{quasipoisson regression fitted by \link{glm} with 
\code{family = quasipoisson()}}
\item{\code{negbin}: }{negative binomial regression fitted by \link[MASS]{glm.nb}}
\item{\code{multinomial}: }{multinomial regression fitted by \link[nnet]{multinom}}
\item{\code{ordinal}: }{ordered logistic regression fitted by \link[MASS]{polr}}
\item{\code{coxph}: }{cox proportional hazard model fitted by \link[survival]{coxph}}
\item{\code{aft_exp}: }{accelerated failure time model fitted by \link[survival]{survreg}
with \code{dist = "exponential"}}
\item{\code{aft_weibull}: }{accelerated failure time model fitted by \link[survival]{survreg}
with \code{dist = "weibull"}}
}
\code{coxph}, \code{aft_exp} and \code{aft_weibull} are currently not implemented for \code{mreg}.

If \code{EMint} is \code{TRUE} and \code{yreg} is specified by the character name of a regression 
model, \code{yreg} is fitted with the interaction between the exposure and each mediator. 


\emph{A Fitted Regression Object:} 
\itemize{
\item{}{Regression objects can be fitted by \link{lm}, \link{glm}, \link{glm.nb}, 
\link[mgcv]{gam}, \link[nnet]{multinom}, \link[MASS]{polr}, \link[survival]{coxph} and
\link[survival]{survreg}. }
\item{}{Regression objects fitted by \link[survival]{coxph} and \link[survival]{survreg} 
are currently not supported for \code{mreg}. }
\item{}{\code{yreg} should regress \code{outcome} on \code{exposure},
\code{mediator} and \code{basec}.}
\item{}{For \code{p=1,...,k}, \code{mreg[p]} should regress \code{mediator[p]} on 
\code{exposure} and \code{basec}, where \code{k} is the number of mediators.}
\item{}{\code{yreg} can't include mediator-mediator interactions when there 
are multiple mediators (VanderWeele TJ & Vansteelandt, 2014).}
}


\strong{Estimation Methods} 
\itemize{
\item{\code{paramfunc}: }{(only available for a single 
mediator) \emph{closed-form parameter function estimation} by Valeri & VanderWeele (2013).
Each causal effect is estimated by a closed-form formula of regression coefficients. }
\item{\code{imputation}: }{\emph{direct counterfactual imputation estimation} by Imai, et al (2010). 
Each causal effect is estimated by imputing counterfactuals directly.}
}

To use \code{paramfunc}, \code{yreg} and \code{mreg} must be specified by the character name 
of a regression model. \code{yreg} can be chosen from \code{linear}, \code{logistic}, \code{loglinear}, 
\code{poisson}, \code{quasipoisson}, \code{negbin}, \code{coxph}, \code{aft_exp} and 
\code{aft_weibull}. \code{mreg} can be chosen from \code{linear}, \code{logistic} and 
\code{multinomial}.

To use \code{paramfunc} with \code{yreg = "logistic"} or \code{yreg = "coxph"}, the outcome must 
be rare.


\strong{Inference Methods}
\itemize{
\item{\code{delta}: }{(only available when \code{estimation = "paramfunc"}) inferences 
about causal effects are obtained by the delta method. }
\item{\code{bootstrap}: }{inferences about causal effects are obtained by bootstrapping. }
}  


\strong{Estimated Causal Effects}

For a continuous outcome, causal effects on the difference scale are estimated. For a 
categorical, count or survival outcome, causal effects on the ratio scale are estimated. Depending 
on the outcome type, the ratio can be risk ratio for a categorical outcome, rate ratio for a count outcome, 
hazard ratio for a survival outcome fitted by \link[survival]{coxph}, mean survival ratio for a survival outcome fitted 
by \link[survival]{survreg}, etc.

When \code{EMint} is \code{FALSE}, \emph{two-way decomposition} (Valeri & VanderWeele, 2013) is conducted, i.e.,
\itemize{
\item{for a continuous outcome: }{\code{cde} (controlled direct effect), \code{pnde} (pure natural 
direct effect), \code{tnde} (total natural direct effect), \code{pnie} (pure natural indirect 
effect), \code{tnie} (total natural indirect effect), \code{te} (total effect), and 
\code{pm} (proportion mediated) are estimated. }
\item{for a categorical, count or survival outcome: }{\code{Rcde} (\code{cde} ratio), \code{Rpnde} (\code{pnde} ratio), 
\code{Rtnde} (\code{tnde} ratio), \code{Rpnie} (\code{pnie} ratio), \code{Rtnie} (\code{tnie} ratio), 
\code{Rte} (\code{te} ratio), and \code{pm} are estimated.}
}

When \code{EMint} is \code{TRUE}: additional \emph{four-way decomposition} (VanderWeele, 2014) is conducted, i.e., 
\itemize{
\item{for a continuous outcome: }{ \code{intref} 
(reference interaction), \code{intmed} (mediated interaction), 
\code{cde(prop)} (proportion \code{cde}), \code{intref(prop)} (proportion 
\code{intref}), \code{intmed(prop)} (proportion \code{intmed}), \code{pnie(prop)} 
(proportion \code{pnie}), \code{int} (proportion 
attributable to interaction), and \code{pe} (proportion eliminated) are estimated.}
\item{for a categorical, count or survival outcome: }{\code{ERcde} (excess ratio due to \code{cde}), \code{ERintref} (excess 
ratio due to \code{intref}), \code{ERintmed} (excess ratio due to \code{intmed}), \code{ERpnie} 
(excess ratio due to \code{pnie}), \code{ERcde(prop)} (proportion \code{ERcde}), 
\code{ERintref(prop)} (proportion \code{ERintref}), \code{ERintmed(prop)} (proportion \code{ERintmed}), 
\code{ERpnie(prop)} (proportion \code{ERpnie}), \code{int}, and \code{pe} are estimated. }
}

When \code{EMint} is \code{TRUE} and \code{estimation} is \code{paramfunc}, 
causal effects conditional on \code{basecval} are estimated. 
Otherwise, marginal causal effects are estimated.
}
\examples{

\dontrun{
library(CMAverse)

# single-mediator case without exposure-mediator interaction
exp1 <- cmest_rb(data = cma2020, outcome = "contY", 
exposure = "A", mediator = "M1", basec = c("C1", "C2"), 
EMint = FALSE, yreg = "linear", mreg = list("logistic"), 
estimation = "paramfunc", inference = "delta", astar = 0, a = 1, mval = list(1))
summary(exp1)

# single-mediator case with exposure-mediator interaction
exp2 <- cmest_rb(data = cma2020, outcome = "contY", 
exposure = "A", mediator = "M2", basec = c("C1", "C2"), 
EMint = TRUE, yreg = "linear", mreg = list("multinomial"), 
estimation = "paramfunc", inference = "delta", astar = 0, a = 1, mval = list("M2_0"))
summary(exp2)

# multiple-mediators case
exp3 <- cmest_rb(data = cma2020, outcome = "contY", 
exposure = "A", mediator = c("M1", "M2"), EMint = TRUE, basec = c("C1", "C2"), 
yreg = "linear", mreg = list("logistic", "multinomial"), 
estimation = "imputation", inference = "bootstrap", 
astar = 0, a = 1, mval = list(0, "M2_0"), 
nboot = 100, boot.ci.type = "bca")
summary(exp3)

# specify regression models by fitted regression objects
exp4 <- cmest_rb(data = cma2020, outcome = "contY", 
exposure = "A", mediator = c("M1", "M2"), EMint = TRUE, basec = c("C1", "C2"), 
yreg = lm(contY ~ A + M1 + M2 + C1 + C2, data = cma2020), 
mreg = list(glm(M1 ~ A + C1 + C2, data = cma2020, family = binomial()),
nnet::multinom(M2 ~ A + C1 + C2, data = cma2020)),
estimation = "imputation", inference = "bootstrap", 
astar = 0, a = 1, mval = list(0, "M2_0"), 
nboot = 100, boot.ci.type = "bca")
summary(exp4)

}

}
\references{
Valeri L, VanderWeele TJ (2013). Mediation analysis allowing for
exposure-mediator interactions and causal interpretation: theoretical assumptions and
implementation with SAS and SPSS macros. Psychological Methods. 18(2): 137 - 150.

VanderWeele TJ, Vansteelandt S (2014). Mediation analysis with multiple mediators.
Epidemiologic Methods. 2(1): 95 - 115.
 
VanderWeele TJ (2014). A unification of mediation and interaction: a 4-way decomposition. 
Epidemiology. 25(5): 749 - 61.

Imai K, Keele L, Tingley D (2010). A general approach to causal mediation analysis.
Psychological Methods. 15(4): 309 - 334.

Schomaker M, Heumann C (2018). Bootstrap inference when using multiple 
imputation. Statistics in Medicine. 37(14): 2252 - 2266. 

Efron B (1987). Better Bootstrap Confidence Intervals. Journal of the American Statistical 
Association. 82(397): 171-185.
}
\seealso{
\link{cmest_gformula}, \link{cmest_wb}, \link{cmest_iorw}, \link{cmest_msm}, 
\link{cmest_multistate}, \link{ggcmest}, \link{cmdag}, \link{cmsens}.
}
