% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmsens.R
\name{cmsens}
\alias{cmsens}
\alias{print.cmsens.uc}
\alias{print.cmsens.me}
\alias{summary.cmsens.me}
\alias{print.summary.cmsens.me}
\title{Sensitivity Analysis For Unmeasured Confounding and Measurement Error}
\usage{
cmsens(
  object = NULL,
  sens = "uc",
  MEmethod = "simex",
  MEvariable = NULL,
  MEvartype = NULL,
  MEerror = NULL,
  lambda = c(0.5, 1, 1.5, 2),
  B = 200,
  nboot.rc = 400
)

\method{print}{cmsens.uc}(x, ...)

\method{print}{cmsens.me}(x, ...)

\method{summary}{cmsens.me}(object, ...)

\method{print}{summary.cmsens.me}(x, digits = 4, ...)
}
\arguments{
\item{object}{an object of class \code{cmsens}.}

\item{sens}{sensitivity analysis for unmeasured confounding or measurement error. 
\code{uc} represents unmeasured confounding and \code{me} represents measurement error.
See \code{Details}.}

\item{MEmethod}{method for measurement error correction. \code{rc} represents regression
calibration and \code{simex} represents SIMEX. See \code{Details}.}

\item{MEvariable}{variable measured with error.}

\item{MEvartype}{type of the variable measured with error. Can be \code{continuous} or 
\code{categorical} (first 3 letters are enough).}

\item{MEerror}{a vector of standard deviations of the measurement error (when \code{MEvartype}
is \code{continuous}) or a list of misclassification matrices (when \code{MEvartype}
is \code{categorical}).}

\item{lambda}{a vector of lambdas for SIMEX. Default is \code{c(0.5, 1, 1.5, 2)}.}

\item{B}{number of simulations for SIMEX. Default is \code{200}.}

\item{nboot.rc}{number of boots for estimating the var-cov matrix of coefficients 
with regression calibration. Default is \code{400}.}

\item{x}{an object of class \code{cmsens}}

\item{...}{other arguments.}

\item{digits}{minimal number of significant digits. See \link{print.default}.}
}
\value{
If \code{sens} is \code{uc}, an object of class 'cmsens.uc' is returned:
\item{call}{the function call,}
\item{evalues}{the data frame in which the first three columns are the point estimates, 
lower limits of confidence intervals, upper limits of confidence intervals of causal effects on 
the ratio scale and the last three columns are the E-values on the ratio scale for them,}
If \code{sens} is \code{me}, an object of class 'cmsens.me' is returned:
\item{call}{the function call,}
\item{ME}{a list which might contain MEvariable, MEvartype, MEerror, reliability ratio, lambda and B,}
\item{naive}{naive causal mediation analysis results}
\item{sens}{a list of causal mediation analysis results after correcting errors in 
\code{MEerror},}

...
}
\description{
\code{cmsens} is used to conduct sensitivity analysis for unmeasured confounding via 
the \emph{E-value} approach by Vanderweele et al. (2017) and Smith et al. (2019), and 
sensitivity analysis for measurement error via \emph{regression calibration} by Carroll 
et al. (1995) and \emph{SIMEX} by Cook et al. (1994) and Küchenhoff et al. (2006).
}
\details{
For unmeasured confounding, all E-values are on the (risk or rate) ratio scale. If the causal 
effects are estimated on the difference scale, the point estimates are transformed into 
ratios using the transformation described by Vanderweele et al. (2017).

Sensitivity analysis for measurement error only supports a single variable measured with error.
Regression calibration only supports an independent continuous variable in a regression 
formula measured with error. SIMEX supports a continuous or categorical variable measured 
with error.
}
\section{Methods (by generic)}{
\itemize{
\item \code{print}: Print the results of cmsens.uc nicely

\item \code{print}: Print the results of cmsens.me nicely

\item \code{summary}: Summarize the results of cmsens.me nicely

\item \code{print}: Print the summary of cmsens.me nicely
}}

\examples{

\dontrun{
library(CMAverse)

# 10 boots are used for illustration
naive <- cmest(data = cma2020, model = "rb", outcome = "contY", 
exposure = "A", mediator = c("M1", "M2"), 
basec = c("C1", "C2"), EMint = TRUE,
mreg = list("logistic", "multinomial"), yreg = "linear",
astar = 0, a = 1, mval = list(0, "M2_0"),
estimation = "imputation", inference = "bootstrap", nboot = 10)

exp1 <- cmsens(object = naive, sens = "uc")
exp2 <- cmsens(object = naive, sens = "me", MEmethod = "rc", 
MEvariable = "C1", MEvartype = "con",
MEerror = c(0.1, 0.2))
summary(exp2)

# B = 10 is used for illustration
exp3 <- cmsens(object = naive, sens = "me", MEmethod = "simex", 
MEvariable = "M1", MEvartype = "cat",
MEerror = list(matrix(c(0.95,0.05,0.05,0.95), nrow = 2), 
matrix(c(0.9,0.1,0.1,0.9), nrow = 2)), B = 10)
summary(exp3)
}

}
\references{
VanderWeele TJ, Ding P. Sensitivity analysis in observational research: introducing the 
E-Value (2017). Annals of Internal Medicine. 167(4): 268 - 274.

Smith LH, VanderWeele TJ. Mediational E-values: Approximate sensitivity analysis for 
unmeasured mediator-outcome confounding (2019). Epidemiology. 30(6): 835 - 837.

Carrol RJ, Ruppert D, Stefanski LA, Crainiceanu C. Measurement Error in Nonlinear Models: 
A Modern Perspective, Second Edition (2006). London: Chapman & Hall.

Cook JR, Stefanski LA. Simulation-extrapolation estimation in parametric measurement error 
models (1994). Journal of the American Statistical Association, 89(428): 1314 - 1328.

Küchenhoff H, Mwalili SM, Lesaffre E. A general method for dealing with misclassification 
in regression: the misclassification SIMEX (2006). Biometrics. 62(1): 85 - 96.

Stefanski LA, Cook JR. Simulation-extrapolation: the measurement error jackknife (1995). 
Journal of the American Statistical Association. 90(432): 1247 - 56.

Valeri L, Lin X, VanderWeele TJ. Mediation analysis when a continuous mediator is measured 
with error and the outcome follows a generalized linear model (2014). Statistics in 
medicine, 33(28): 4875–4890.
}
\seealso{
\code{\link{ggcmsens}}, \code{\link{cmdag}}, \code{\link{cmest}}
}
