% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmsens.R
\name{ggcmsens}
\alias{ggcmsens}
\title{Plotting Results of Sensitivity Analysis for Measurement Error}
\usage{
ggcmsens(
  x,
  errorbar.width = 0.3,
  errorbar.size = 0.3,
  errorbar.position = position_dodge2(width = 0.5),
  point.size = 1,
  point.position = position_dodge2(width = 0.3),
  refline = TRUE,
  refline.colour = "red",
  refline.size = 0.3
)
}
\arguments{
\item{x}{an object of class \code{cmsens.me}}

\item{errorbar.width}{width of errorbars for confidence intervals. Default is \code{0.3}.}

\item{errorbar.size}{size of errorbars for confidence intervals. Default is \code{0.3}.}

\item{errorbar.position}{position adjustment for confidence intervals, either as a string, 
or the result of a call to a position adjustment function. Default is 
\code{position_dodge2(width = 0.5)}. See \link[ggplot2]{geom_errorbar} for details.}

\item{point.size}{size of points for point estimates. Default is \code{1}.}

\item{point.position}{position adjustment for point estimates, either as a string, or 
the result of a call to a position adjustment function. Default is 
\code{position_dodge2(width = 0.3)}. See \link[ggplot2]{geom_errorbar} for details.}

\item{refline}{a logical value. If \code{true}, include a reference line at 
\code{y = 0} when effects are on the difference scale and include a reference line at 
\code{y = 1} when effects are on the ratio scale. Default is \code{TRUE}.}

\item{refline.colour}{colour of the reference line. Default is \code{red}.}

\item{refline.size}{size of the reference line. Default is \code{0.3}.}
}
\description{
\code{ggcmsens} is used to plot results of \code{cmsens} nicely with plotting functions
in the \code{ggplot2} package. Additional layers can be added to this plot using other 
plotting functions in the \code{ggplot2} package.
}
\examples{

library(CMAverse)
library(ggplot2)

naive <- cmest(data = cma2020, model = "rb", outcome = "contY", 
exposure = "A", mediator = c("M1", "M2"), 
basec = c("C1", "C2"), EMint = TRUE,
mreg = list("logistic", "multinomial"), yreg = "linear",
astar = 0, a = 1, mval = list(0, "M2_0"),
estimation = "imputation", inference = "bootstrap", nboot = 10)

x <- cmsens(object = naive, sens = "me", MEmethod = "rc", 
MEvariable = "C1", MEvartype = "con",
MEerror = c(0.1, 0.2))

ggcmsens(x) +
ggplot2::theme(axis.text.x = ggplot2::element_text(angle = 45))

ggcmsens(x) +
ggplot2::coord_flip(xlim = NULL, ylim = NULL, expand = TRUE, clip = "on")

}
\seealso{
\code{\link{cmsens}}, \code{\link{ggplot2}}.
}
