% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simexreg.R
\name{simexreg}
\alias{simexreg}
\alias{coef.simexreg}
\alias{vcov.simexreg}
\alias{sigma.simexreg}
\alias{formula.simexreg}
\alias{family.simexreg}
\alias{predict.simexreg}
\alias{model.frame.simexreg}
\alias{print.simexreg}
\alias{summary.simexreg}
\alias{print.summary.simexreg}
\alias{update.simexreg}
\title{Simulation and Extrapolation for Measurement Error Correction}
\usage{
simexreg(
  reg = NULL,
  data = NULL,
  weights = NULL,
  MEvariable = NULL,
  MEvartype = NULL,
  MEerror = NULL,
  variance = FALSE,
  lambda = c(0.5, 1, 1.5, 2),
  B = 200
)

\method{coef}{simexreg}(object, ...)

\method{vcov}{simexreg}(object, ...)

\method{sigma}{simexreg}(object, ...)

\method{formula}{simexreg}(x, ...)

\method{family}{simexreg}(object, ...)

\method{predict}{simexreg}(object, ...)

\method{model.frame}{simexreg}(formula, ...)

\method{print}{simexreg}(x, ...)

\method{summary}{simexreg}(object, ...)

\method{print}{summary.simexreg}(x, digits = 4, ...)

\method{update}{simexreg}(object, ..., evaluate = TRUE)
}
\arguments{
\item{reg}{naive regression object. See \code{Details}.}

\item{data}{new dataset for \code{reg}}

\item{weights}{new weights for \code{reg}}

\item{MEvariable}{variable measured with error}

\item{MEvartype}{type of the variable measured with error. Can be \code{continuous} or 
\code{categorical} (first 3 letters are enough).}

\item{MEerror}{the standard deviation of the measurement error (when \code{MEvartype}
is \code{continuous}) or the misclassification matrix (when \code{MEvartype}
is \code{categorical}).}

\item{variance}{a logical value. If \code{TRUE}, estimate the var-cov matrix of
coefficients through Jackknife. Default is \code{FALSE}.}

\item{lambda}{a vector of lambdas for \emph{SIMEX}. Default is \code{c(0.5, 1, 1.5, 2)}.}

\item{B}{number of simulations for \emph{SIMEX}. Default is \code{200}.}

\item{object}{an object of class \code{simexreg}}

\item{...}{additional arguments}

\item{x}{an object of class \code{simexreg}}

\item{formula}{an object of class \code{simexreg}}

\item{digits}{minimal number of significant digits. See \link{print.default}.}

\item{evaluate}{a logical value. If \code{TRUE}, the updated call is evaluated. Default
is \code{TRUE}.}
}
\value{
If \code{MEvariable} is not in the regression formula of \code{reg}, \code{reg} is 
returned. If \code{MEvariable} is in the regression formula of \code{yeg}, an object of 
class \code{simexreg} is returned:
\item{call}{the function call,}
\item{NAIVEreg}{the naive regression object,}
\item{ME}{a list of \code{MEvariable}, \code{MEvartype}, \code{MEerror}, \code{variance},
\code{lambda} and \code{B},}
\item{RCcoef}{coefficient estimates corrected by SIMEX,}
\item{RCsigma}{the residual standard deviation of a linear regression object corrected by 
SIMEX,}
\item{RCvcov}{the var-cov matrix of coefficients corrected by SIMEX,}
...
}
\description{
\code{simexreg} is used to correct a regression object with a variable measured with 
error via \emph{SIMEX} by Cook et al. (1994) and Küchenhoff et al. (2006).
}
\details{
\code{reg} fitted by \link{lm}, \link{glm} (with family \code{gaussian}, \code{binomial} or
\code{poisson}), \link[nnet]{multinom}, \link[MASS]{polr}, \link[survival]{coxph} or
\link[survival]{survreg} is supported.
}
\section{Methods (by generic)}{
\itemize{
\item \code{coef}: Extract coefficients corrected by \code{simexreg}

\item \code{vcov}: Extract the var-cov matrix of coefficients corrected by 
\code{simexreg}

\item \code{sigma}: Extract the residual standard deviation of a linear regression object 
corrected by \code{simexreg}

\item \code{formula}: Extract the regression formula

\item \code{family}: Extract the family of a regression of class \code{lm} or \code{glm}

\item \code{predict}: Predict with new data

\item \code{model.frame}: Extract the model frame

\item \code{print}: Print results of \code{simexreg} nicely

\item \code{summary}: Summarize results of \code{simexreg} nicely

\item \code{print}: Print summary of \code{simexreg} nicely

\item \code{update}: Update \code{simexreg}
}}

\examples{

\dontrun{
rm(list=ls())
library(CMAverse)

# lm
n <- 1000
x1 <- rnorm(n, mean = 5, sd = 3)
x2_true <- rnorm(n, mean = 2, sd = 1)
error1 <- rnorm(n, mean = 0, sd = 0.5)
x2_error <- x2_true + error1
x3 <- rbinom(n, size = 1, prob = 0.4)
y <- 1 + 2 * x1 + 4 * x2_true + 2 * x3  + rnorm(n, mean = 0, sd = 2)
data <- data.frame(x1 = x1, x2_true = x2_true, x2_error = x2_error,
                   x3 = x3, y = y)
reg_naive <- lm(y ~ x1 + x2_error + x3, data = data)
reg_true <- lm(y ~ x1 + x2_true + x3, data = data)
reg_simex <- simexreg(reg = reg_naive, data = data, MEvariable = "x2_error", 
MEvartype = "con", MEerror = 0.5, variance = TRUE)
coef(reg_simex)
vcov(reg_simex)
sigma(reg_simex)
formula(reg_simex)
family(reg_simex)
predict(reg_simex, newdata = data[1, ])
reg_simex_model <- model.frame(reg_simex)
reg_simex_update <- update(reg_simex, data = data, weights = rep(1, n))
reg_simex_summ <- summary(reg_simex)
                
# glm
n <- 1000
x1 <- rnorm(n, mean = 5, sd = 3)
x2_true <- sample(x = c(1:3), size = n, prob = c(0.2,0.3,0.5), replace = TRUE)
MEerror <- matrix(c(0.8,0.1,0.1,0.2,0.7,0.1,0.05,0.25,0.7), nrow = 3)
x2_error <- x2_true
for (j in 1:3) {
  x2_error[which(x2_error == c(1:3)[j])] <-
    sample(x = c(1:3), size = length(which(x2_error == c(1:3)[j])),
           prob = MEerror[, j], replace = TRUE)
}
x2_true <- as.factor(x2_true)
x2_error <- as.factor(x2_error)
x3 <- rnorm(n, mean = 2, sd = 1)
linearpred <- 1 + 0.3 * x1 - 1.5*(x2_true == 2) - 2.5*(x2_true == 3) - 0.2 * x3
py <- exp(linearpred) / (1 + exp(linearpred))
y <- rbinom(n, size = 1, prob = py)
data <- data.frame(x1 = x1, x2_true = x2_true, x2_error = x2_error,
                   x3 = x3, y = y)
reg_naive <- glm(y ~ x1 + x2_error + x3, data = data, family = binomial("logit"))
reg_true <- glm(y ~ x1 + x2_true + x3, data = data, family = binomial("logit"))
reg_simex <- simexreg(reg = reg_naive, data = data, MEvariable = "x2_error",
                      MEerror = MEerror, variance = TRUE, MEvartype = "cat")
}
                  
}
\references{
Carrol RJ, Ruppert D, Stefanski LA, Crainiceanu C (2006). Measurement Error in Nonlinear Models: 
A Modern Perspective, Second Edition. London: Chapman & Hall.

Cook JR, Stefanski LA (1994). Simulation-extrapolation estimation in parametric measurement error 
models. Journal of the American Statistical Association, 89(428): 1314 - 1328.

Küchenhoff H, Mwalili SM, Lesaffre E (2006). A general method for dealing with misclassification 
in regression: the misclassification SIMEX. Biometrics. 62(1): 85 - 96.

Stefanski LA, Cook JR (1995). Simulation-extrapolation: the measurement error jackknife. 
Journal of the American Statistical Association. 90(432): 1247 - 56.
}
\seealso{
\code{\link{rcreg}}, \code{\link{cmsens}}, \code{\link{cmest}}.
}
