% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svymultinom.R
\name{svymultinom}
\alias{svymultinom}
\alias{coef.svymultinom}
\alias{vcov.svymultinom}
\alias{formula.svymultinom}
\alias{predict.svymultinom}
\alias{model.frame.svymultinom}
\alias{print.svymultinom}
\alias{summary.svymultinom}
\alias{print.summary.svymultinom}
\alias{update.svymultinom}
\title{Multinomial Regression with Survey Data}
\usage{
svymultinom(formula = NULL, weights = NULL, data = NULL)

\method{coef}{svymultinom}(object, ...)

\method{vcov}{svymultinom}(object, ...)

\method{formula}{svymultinom}(x, ...)

\method{predict}{svymultinom}(object, ...)

\method{model.frame}{svymultinom}(formula, ...)

\method{print}{svymultinom}(x, ...)

\method{summary}{svymultinom}(object, ...)

\method{print}{summary.svymultinom}(x, digits = 4, ...)

\method{update}{svymultinom}(object, ..., evaluate = TRUE)
}
\arguments{
\item{formula}{regression formula}

\item{weights}{weights for regression}

\item{data}{dataset for regression}

\item{object}{an object of class 'svymultinom'}

\item{...}{additional arguments}

\item{x}{an object of class 'svymultinom'}

\item{digits}{minimal number of significant digits. See \link{print.default}.}

\item{evaluate}{a logical value. If \code{TRUE}, the updated call is evaluated. Default
is \code{TRUE}.}
}
\value{
An object of class \code{svymultinom} is returned:
\item{call}{the function call,}
\item{NAIVEreg}{the naive multinomial regression object,}
\item{vcov}{the replicate-based estimate of the variance-covariance matrix of coefficients,}
...
}
\description{
\code{svymultinom} uses the \link[survey]{withReplicates} function to compute the replicate-based 
estimate of the variance-covariance matrix of coefficients for a multinomial regression fitted by 
\link[nnet]{multinom}.
}
\section{Methods (by generic)}{
\itemize{
\item \code{coef}: Extract coefficients

\item \code{vcov}: Extract the var-cov matrix of coefficients

\item \code{formula}: Extract the regression formula

\item \code{predict}: Predict with new data

\item \code{model.frame}: Extract the model frame

\item \code{print}: Print results of \code{svymultinom} nicely

\item \code{summary}: Summarize results of \code{svymultinom} nicely

\item \code{print}: Print summary of \code{svymultinom} nicely

\item \code{update}: Update \code{svymultinom}
}}

\examples{

\dontrun{
rm(list=ls())
library(CMAverse)

# multinom
n <- 1000
x1 <- rnorm(n, mean = 0, sd = 1)
x2 <- rnorm(n, mean = 1, sd = 1)
x3 <- rbinom(n, size = 1, prob = 0.4)
linearpred1 <- 1 + 0.3 * x1 - 0.5 * x2 - 0.2 * x3
linearpred2 <- 2 + 1 * x1 - 2 * x2 - 1 * x3
py2 <- exp(linearpred1) / (1 + exp(linearpred1) + exp(linearpred2))
py3 <- exp(linearpred2) / (1 + exp(linearpred1) + exp(linearpred2))
py1 <- 1 - py2 - py3
y <- sapply(1:n, function(x) sample(size = 1, c(1:3), prob = c(py1[x], py2[x], py3[x])))
w <- ifelse(x3 == 0, 0.4, 0.6)
data <- data.frame(x1 = x1, x2 = x2, x3 = x3, y = y)

reg <- svymultinom(y ~ x1 + x2 + x3, weights = w, data = data)
coef(reg)
vcov(reg)
formula(reg)
predict(reg, newdata = data[1, ], type = "probs")
model.frame(reg)
summary(reg)
update(reg, weights = w[1:500], data = data[1:500, ])
}
 
}
\seealso{
\link[nnet]{multinom}, \link[survey]{withReplicates}
}
