% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmdag.R
\name{cmdag}
\alias{cmdag}
\title{DAG Visualization}
\usage{
cmdag(
  outcome = NULL,
  exposure = NULL,
  mediator = NULL,
  basec = NULL,
  postc = NULL,
  x.outcome = 4,
  x.exposure = 0,
  x.mediator = 2,
  x.basec = 2,
  x.postc = 2,
  y.outcome = 0,
  y.exposure = 0,
  y.mediator = 1,
  y.basec = 2,
  y.postc = -0.5,
  caption.width = 50,
  caption.size = 10,
  ...
)
}
\arguments{
\item{outcome}{the variable name of the outcome}

\item{exposure}{the variable name of the exposure}

\item{mediator}{a vector of variable name(s) of the mediator(s)}

\item{basec}{(optional) a vector of variable name(s) of the exposure-outcome
confounder(s), exposure-mediator confounder(s) and mediator-outcome 
confounder(s) not affected by the exposure}

\item{postc}{(optional) a vector of variable name(s) of the mediator-outcome 
confounder(s) affected by the exposure}

\item{x.outcome}{x coordinate of \code{outcome}. Default is \code{4}.}

\item{x.exposure}{x coordinate of \code{exposure}. Default is \code{0}.}

\item{x.mediator}{x coordinate of \code{mediator}. Default is \code{2}.}

\item{x.basec}{x coordinate of \code{basec}. Default is \code{2}.}

\item{x.postc}{x coordinate of \code{postc}. Default is \code{2}.}

\item{y.outcome}{y coordinate of \code{outcome}. Default is \code{0}.}

\item{y.exposure}{y coordinate of \code{exposure}. Default is \code{0}.}

\item{y.mediator}{y coordinate of \code{mediator}. Default is \code{1}.}

\item{y.basec}{y coordinate of \code{basec}. Default is \code{2}.}

\item{y.postc}{y coordinate of \code{postc}. Default is \code{-0.5}.}

\item{caption.width}{line width in characters for the caption. Default is \code{50}.}

\item{caption.size}{text size in pts for the caption. Default is \code{10}.}

\item{...}{additional arguments passed to \code{ggdag()}. See \link[ggdag]{ggdag} for details.}
}
\description{
Plot the directed acyclic graph (DAG) for causal mediation analysis.
}
\examples{

## basec and postc are empty
cmdag(outcome = "Y", exposure = "A", mediator = c("M1", "M2"),
     basec = NULL, postc = NULL, node = TRUE, text_col = "white")

## postc is empty
cmdag(outcome = "Y", exposure = "A", mediator = c("M1", "M2"),
     basec = c("C1", "C2", "C3"), postc = NULL, node = FALSE, text_col = "black")
     
## basec is empty
cmdag(outcome = "Y", exposure = "A", mediator = c("M1", "M2"),
     basec = NULL, postc = c("L1", "L2"), node = FALSE, text_col = "black")

## basec and postc aren't empty
cmdag(outcome = "Y", exposure = "A", mediator = c("M1", "M2"),
     basec = c("C1", "C2", "C3"), postc = c("L1", "L2"), node = FALSE, text_col = "black")

}
\seealso{
\link{cmest}, \link{cmsens}.
}
