% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmest.R
\name{cmest}
\alias{cmest}
\alias{print.cmest}
\alias{summary.cmest}
\alias{print.summary.cmest}
\title{Causal Mediation Analysis}
\usage{
cmest(
  data = NULL,
  model = "rb",
  full = TRUE,
  casecontrol = FALSE,
  yrare = NULL,
  yprevalence = NULL,
  estimation = "imputation",
  inference = "bootstrap",
  outcome = NULL,
  event = NULL,
  exposure = NULL,
  mediator = NULL,
  EMint = NULL,
  basec = NULL,
  postc = NULL,
  yreg = NULL,
  mreg = NULL,
  wmnomreg = NULL,
  wmdenomreg = NULL,
  ereg = NULL,
  postcreg = NULL,
  astar = 0,
  a = 1,
  mval = NULL,
  yval = NULL,
  basecval = NULL,
  nboot = 200,
  boot.ci.type = "per",
  nRep = 5,
  multimp = FALSE,
  ...
)

\method{print}{cmest}(x, ...)

\method{summary}{cmest}(object, ...)

\method{print}{summary.cmest}(x, digits = 4, ...)
}
\arguments{
\item{data}{dataset}

\item{model}{causal mediation analysis approach. \code{rb}, \code{wb}, \code{iorw}, 
\code{ne}, \code{msm} and \code{gformula} are implemented. Default is \code{rb}. 
See \code{Details}.}

\item{full}{a logical value. If \code{TRUE}, output a full list of causal effects; if 
\code{FALSE}, output a reduced list of causal effects. Default is \code{TRUE}.}

\item{casecontrol}{a logical value. \code{TRUE} indicates a case control study in which the
first level of the outcome is treated as the control and the second level of the outcome is 
treated as the case. Default is \code{FALSE}.}

\item{yrare}{a logical value (used when \code{casecontrol} is \code{TRUE}). \code{TRUE} 
indicates the case is rare.}

\item{yprevalence}{the prevalence of the case (used when \code{casecontrol} is \code{TRUE}).}

\item{estimation}{method for estimating causal effects. \code{paramfunc} and 
\code{imputation} are implemented (the first 4 letters are enough). Default is \code{imputation}. 
See \code{Details}.}

\item{inference}{method for estimating standard errors of causal effects. \code{delta} and 
\code{bootstrap} are implemented (the first 4 letters are enough). Default is \code{bootstrap}. 
See \code{Details}.}

\item{outcome}{variable name of the outcome.}

\item{event}{variable name of the event (used when \code{yreg} is \code{coxph}, \code{aft_exp},
or \code{aft_weibull}).}

\item{exposure}{variable name of the exposure.}

\item{mediator}{a vector of variable name(s) of the mediator(s) following the temporal order.}

\item{EMint}{a logical value indicating the existence of exposure-mediator interaction in 
\code{yreg} (used when \code{model} is not \code{iorw}). \code{TRUE} when there is 
exposure-mediator interaction in \code{yreg}. If \code{TRUE} and character \code{yreg} is 
provided, the outcome regression includes exposure-mediator interaction(s) between the exposure 
and each of the mediator(s).}

\item{basec}{a vector of variable name(s) of the exposure-outcome confounder(s), exposure-mediator 
confounder(s) and mediator-outcome confounder(s) not affected by the exposure}

\item{postc}{a vector of variable name(s) of the mediator-outcome confounder(s) affected 
by the exposure following the temporal order}

\item{yreg}{outcome regression model. See \code{Details}.}

\item{mreg}{a list specifying a regression model for each variable in \code{mediator} (used 
when \code{model} is \code{rb}, \code{msm} or \code{gformula}). The order of regression models 
must follow the order of variables in \code{mediator}. See \code{Details}.}

\item{wmnomreg}{a list specifying a regression model for calculating the nominators of 
weights with respect to each variable in \code{mediator} (used when \code{model} is \code{msm}). 
The order of regression models must follow the order of variables in \code{mediator}. See 
\code{Details}.}

\item{wmdenomreg}{a list specifying a regression model for calculating the denominators of 
weights with respect to each variable in \code{mediator} (used when \code{model} is \code{msm}). 
The order of regression models must follow the order of variables in \code{mediator}. See 
\code{Details}.}

\item{ereg}{exposure regression model for calculating weights with respect to the exposure (used 
when \code{model} is \code{wb} or \code{msm} with a non-empty \code{basec} or when 
\code{model} is \code{iorw}). See \code{Details}.}

\item{postcreg}{a list specifying a regression model for each variable in \code{postc} (used 
when \code{model} is \code{gformula}). The order of regression models must follow the order of 
variables in \code{postc}. See \code{Details}.}

\item{astar}{the control value for the exposure. Default is \code{0}.}

\item{a}{the active value for the exposure. Default is \code{1}.}

\item{mval}{a list specifying a value for each variable in \code{mediator} at which the variable is 
controlled (used when \code{model} is \code{rb}, \code{wb}, \code{ne}, \code{msm} or \code{gformula}).}

\item{yval}{the value of the outcome at which causal effects on the risk/odds ratio 
scale are estimated (used when the outcome is categorical).}

\item{basecval}{a list specifying a conditional value for each variable in \code{basec} conditional 
on which causal effects are estimated (used when \code{estimation} is \code{paramfunc}). The order 
of conditional values must follow the order of variables in \code{basec}. If \code{NULL}, 
mean values of variable(s) in \code{basec} are used.}

\item{nboot}{the number of boots applied (used when \code{inference} is \code{bootstrap}). 
Default is 200.}

\item{boot.ci.type}{the type of bootstrap confidence interval. If \code{per}, percentile bootstrap
confidence intervals are estimated; if \code{bca}, bias-corrected and accelerated (BCa) bootstrap 
confidence intervals are estimated. Default is \code{per}.}

\item{nRep}{number of replications or hypothetical values of the exposure to sample for 
each observation unit (used when \code{model} is \code{ne}). Default is \code{5}.}

\item{multimp}{a logical value (used when \code{data} contains missing values). If 
\code{TRUE}, conduct multiple imputations using the \link[mice]{mice} function. Default is 
\code{FALSE}.}

\item{...}{Additional arguments passed to the \link[mice]{mice} function. See \link[mice]{mice}
for details.}

\item{x}{an object of class \code{cmest}}

\item{object}{an object of class \code{cmest}}

\item{digits}{minimal number of significant digits. See \link{print.default}.}
}
\value{
An object of class \code{cmest} is returned:
\item{call}{the function call,}
\item{data}{the dataset,}
\item{methods}{a list of methods used which may include \code{model}, \code{full}, 
\code{casecontrol}, \code{yprevalence}, \code{yrare}, \code{estimation}, \code{inference}, 
\code{nboot}, \code{boot.ci.type} and \code{nRep},}
\item{variables}{a list of variables used which may include \code{outcome}, \code{event}, 
\code{exposure}, \code{mediator}, \code{EMint}, \code{basec} and \code{postc},}
\item{reg.input}{a list of regressions input,}
\item{multimp}{a list of arguments used for multiple imputation,}
\item{ref}{reference values used which may include \code{a}, \code{astar}, \code{mval}, 
\code{yval} and \code{basecval},}
\item{reg.output}{a list of regressions output. If \code{multimp} is \code{TRUE}, 
reg.output contains regressions fitted by each of the imputed dataset,}
\item{effect.pe}{point estimates of causal effects,}
\item{effect.se}{standard errors of causal effects,}
\item{effect.ci.low}{the lower limits of 95\% confidence intervals of causal effects,}
\item{effect.ci.high}{the higher limits of 95\% confidence intervals of causal effects,}
\item{effect.pval}{p-values of causal effects,}
...
}
\description{
\code{cmest} is used to implement six causal mediation analysis approaches including 
\emph{the regression-based approach} by Valeri et al. (2013) and VanderWeele 
et al. (2014), \emph{the weighting-based approach} by VanderWeele et al. (2014), 
\emph{the inverse odd-ratio weighting approach} by Tchetgen Tchetgen (2013), 
\emph{the natural effect model} by Vansteelandt et al. (2012), \emph{the marginal structural 
model} by VanderWeele et al. (2017), and \emph{the g-formula approach} by Robins (1986).
}
\details{
\strong{Regressions}

Each regression in \code{yreg}, \code{mreg}, \code{wmnomreg}, \code{wmdenomreg},
\code{ereg} and \code{postcreg} can be specified by a user-defined regression 
object or the character name of the regression. 

\emph{The Character Name of A Regression}

\itemize{
\item{\code{linear}: }{linear regression fitted by \link{glm} with \code{family = gaussian()}}
\item{\code{logistic}: }{logistic regression fitted by \link{glm} with \code{family = logit()}}
\item{\code{loglinear}: }{log linear regression fitted by \link{glm} with 
\code{family = poisson()} for a binary response}
\item{\code{poisson}: }{poisson regression fitted by \link{glm} with 
\code{family = poisson()} for a count response}
\item{\code{quasipoisson}: }{quasipoisson regression fitted by \link{glm} with 
\code{family = quasipoisson()}}
\item{\code{negbin}: }{negative binomial regression fitted by \link[MASS]{glm.nb}}
\item{\code{multinomial}: }{multinomial regression fitted by \link[nnet]{multinom}}
\item{\code{ordinal}: }{ordered logistic regression fitted by \link[MASS]{polr}}
\item{\code{coxph}: }{cox proportional hazard model fitted by \link[survival]{coxph}}
\item{\code{aft_exp}: }{accelerated failure time model fitted by \link[survival]{survreg}
with \code{dist = "exponential"}}
\item{\code{aft_weibull}: }{accelerated failure time model fitted by \link[survival]{survreg}
with \code{dist = "weibull"}}
}
\code{coxph}, \code{aft_exp} and \code{aft_weibull} are currently not implemented for 
\code{mreg}, \code{wmnomreg}, \code{wmdenomreg}, \code{ereg} and \code{postcreg}.

\emph{The User-defined Regression Object} 

A user-defined regression object can be fitted by \link{lm}, \link{glm}, \link{glm.nb}, 
\link[mgcv]{gam}, \link[nnet]{multinom}, \link[MASS]{polr}, \link[survival]{coxph} and
\link[survival]{survreg}. Objects fitted by \link[survival]{coxph} and \link[survival]{survreg} 
are currently not supported for \code{mreg}, \code{wmnomreg}, \code{wmdenomreg}, 
\code{ereg} and \code{postcreg}.

The \code{cmest} function calculates weights for regressions when weighting is required. If a 
user-defined regression object is fitted with prior weights, the final weights for this 
regression object are constructed by multiplying the prior weights and the weights calculated 
inside the \code{cmest} function. 


\strong{Causal Mediation Analysis Approaches}

Let \code{Y} denote the outcome, \code{A} denote the exposure, \code{M=(M_1,...,M_k)^T} 
denote \code{mediator}, \code{C} denote \code{basec}, \code{L=(L_1,...,L_s)^T} denote \code{postc}.

\itemize{
    \item{\code{rb}: }{\emph{the regression-based approach} by Valeri et al. (2013) and
      VanderWeele et al. (2014). \code{yreg} and \code{mreg} are required. If specified as
      a user-defined regression object, \code{yreg} should regress \code{Y} on \code{A},
\code{M} and \code{C} and \code{mreg[p]} should regress \code{M_p} on \code{A} and \code{C}} 
for \code{p=1,...,k}.

    \item{\code{wb}: }{\emph{the weighting-based approach} by VanderWeele et al. (2014).
    \code{yreg} is required. When \code{basec} is not empty, \code{ereg} is also required 
    and \code{A} must be categorical. If specified as a user-defined regression object, 
    \code{yreg} should regress \code{Y} on \code{A}, \code{M} and \code{C} and \code{ereg} 
    should regress \code{A} on \code{C}.}
    
    \item{\code{iorw}: }{\emph{the inverse odd-ratio weighting approach} by
      Tchetgen Tchetgen (2013). \code{yreg} and \code{ereg} are required and 
      \code{A} must be categorical. If specified as a user-defined regression object, 
      \code{yreg} should regress \code{Y} on \code{A} and \code{C} and \code{ereg} should 
      regress \code{A} on \code{M} and \code{C}.}
      
    \item{\code{ne}: }{\emph{the natural effect model} by Vansteelandt et al. (2012). 
    \code{yreg} is required. If specified as a user-defined regression object, \code{yreg}
    should regress \code{Y} on \code{A}, \code{M} and \code{C}. The variables in the 
    formula of \code{yreg} must follow the order of \code{A}, \code{M} and \code{C}, i.e., 
    the first variable must point to the exposure, the variable(s) right after the 
    exposure must point to the mediator(s), e.g., \code{Y ~ A + M_1 + M_2 + A*M_1 + C}.}
      
    \item{\code{msm}: }{\emph{the marginal structural model} by VanderWeele et al. (2017).
    \code{yreg}, \code{mreg}, \code{wmnomreg} and \code{wmdenomreg} are required and all 
    mediators must be categorical. When \code{basec} is not empty, \code{ereg} is also 
    required and \code{A} must be categorical. If specified as a user-defined regression 
    object, \code{yreg} should regress \code{Y} on \code{A} and \code{M}; \code{mreg[p]} 
    should regress \code{M_p} on \code{A} for \code{p=1,...,k}; \code{wmnomreg[p]} should 
    regress \code{M_p} on \code{A}, \code{M_1}, ..., \code{M_{p-1}} for \code{p=1,...,k};
    \code{wmdenomreg[p]} should regress \code{M_p} on \code{A}, \code{M_1}, ..., \code{M_{p-1}},
    \code{C} and \code{L} for \code{p=1,...,k}; and \code{ereg} should regress \code{A} on 
    \code{C}.}
    
    \item{\code{gformula}: }{\emph{the g-formula approach} by Robins (1986). 
    \code{yreg}, \code{mreg} are required. \code{postcreg} is also required when \code{postc}
    is not empty. If specified as a user-defined regression object, \code{yreg} should 
    regress \code{Y} on \code{A}, \code{M}, \code{C} and \code{L}, \code{mreg[p]} should 
    regress \code{M_p} on \code{A}, \code{C} and \code{L}
    for \code{p=1,...,k}, \code{postcreg[q]} should regress \code{L_q} on \code{A} and \code{C} 
    for \code{q=1,...,s}.}
  }
  
When \code{postc} is not empty, only \code{msm} and \code{gformula} can be used.

When there are mediatior-mediator interactions in \code{yreg}, only \code{wb}, \code{iorw},
\code{ne} and \code{msm} can be used.
  

\strong{Estimation Methods} 
  
\itemize{
\item{\code{paramfunc}: }{closed-form parameter function estimation (only available when 
\code{model = "rb"} and \code{length(mediator) = 1}). The point estimate of each causal 
effect is obtained by a closed-form formula of regression coefficients. Effects conditional 
on \code{basecval} are estimated.}
\item{\code{imputation}: }{direct counterfactual imputation estimation. The point estimate 
of each causal effect is obtained by imputing counterfactuals directly.}
}

To use \code{paramfunc}, \code{yreg} and \code{mreg} must be specified by the character name 
of the regression. \code{yreg} can be chosen from \code{linear}, \code{logistic}, \code{loglinear}, 
\code{poisson}, \code{quasipoisson}, \code{negbin}, \code{coxph}, \code{aft_exp} and 
\code{aft_weibull}. \code{mreg} can be chosen from \code{linear}, \code{logistic} and 
\code{multinomial}.

To use \code{paramfunc} with \code{yreg = "logistic"}, the outcome must be rare. To use 
\code{paramfunc} with \code{yreg = "coxph"}, the outcome at the end of follow-up must be rare.


\strong{Inference Methods}

\itemize{
\item{\code{delta}: }{delta method (only available when \code{estimation = "paramfunc"}). 
The standard errors of causal effects are obtained by the delta method. The confidence 
intervals of causal effects are obtained by normal distribution approximation.}
\item{\code{bootstrap}: }{bootstrapping. The standard errors of causal effects are 
obtained by the standard deviations of bootstrapped results. The confidence intervals of 
causal effects are obtained by percentiles of bootstrapped results.}
}  


\strong{Estimated Causal Effects}

For a continuous outcome, the causal effects on the difference scale are estimated. For a 
categorical, count or survival outcome, the causal effects on the ratio scale are estimated. The 
interpretation of the ratio depends on the type of the outcome and it can be risk 
ratio for a categorical outcome, rate ratio for a count outcome, hazard ratio for a survival 
outcome fitted by \link[survival]{coxph}, mean survival ratio for a survival outcome fitted 
by \link[survival]{survreg}, etc.

\emph{Continuous Outcome}

When \code{model = "rb", "wb", "ne", "msm" or "gformula"} with an empty \code{postc} and 
\code{EMint} is \code{TRUE}, \code{cde} (controlled direct effect), \code{pnde} (pure natural 
direct effect), \code{tnde} (total natural direct effect), \code{pnie} (pure natural indirect 
effect), \code{tnie} (total natural indirect effect), \code{te} (total effect),  \code{intref} 
(reference interaction), \code{intmed} (mediated interaction), 
\code{cde(prop)} (proportion \code{cde}), \code{intref(prop)} (proportion 
\code{intref}), \code{intmed(prop)} (proportion \code{intmed}), \code{pnie(prop)} 
(proportion \code{pnie}), \code{pm} (proportion mediated), \code{int} (proportion 
attributable to interaction) and \code{pe} (proportion eliminated) are estimated.

When \code{model = "rb", "wb", "ne", "msm" or "gformula"} with an empty \code{postc} and 
\code{EMint} is \code{FALSE}, \code{cde}, \code{pnde}, \code{tnde}, \code{pnie}, \code{tnie}, 
\code{te} and \code{pm} are estimated.

When \code{postc} is not empty, \code{pnde}, \code{tnde}, \code{pnie}, \code{tnie},  
\code{intref}, \code{intmed}, \code{intref(prop)}, \code{intmed(prop)}, \code{pnie(prop)}, 
\code{pm}, \code{int} and \code{pe} are replaced by their randomized analogues \code{rpnde}, 
\code{rtnde}, \code{rpnie}, \code{rtnie}, \code{rintref}, \code{rintmed}, \code{rintref(prop)}, 
\code{rintmed(prop)}, \code{rpnie(prop)}, \code{rpm}, \code{rint} and \code{rpe}.

When \code{model = "iorw"}, \code{te}, \code{pnde}, \code{tnie} and \code{pm} are estimated. 

\emph{Categorical, Count or Survival Outcome}

When \code{model = "rb", "wb", "ne", "msm" or "gformula"} with an empty \code{postc} and
\code{EMint} is \code{TRUE}, \code{Rcde} (\code{cde} ratio), \code{Rpnde} (\code{pnde} ratio), 
\code{Rtnde} (\code{tnde} ratio), \code{Rpnie} (\code{Rpnie} ratio), \code{Rtnie} (\code{tnie} ratio), 
\code{Rte} (\code{te} ratio), \code{ERcde} (excess ratio due to \code{cde}), \code{ERintref} (excess 
ratio due to \code{intref}), \code{ERintmed} (excess ratio due to \code{intmed}), \code{ERpnie} 
(excess ratio due to \code{pnie}), \code{ERcde(prop)} (proportion \code{ERcde}), 
\code{ERintref(prop)} (proportion \code{ERintref}), \code{ERintmed(prop)} (proportion \code{ERintmed}), 
\code{ERpnie(prop)} (proportion \code{ERpnie}), \code{pm}, \code{int} and \code{pe} are estimated. 

When \code{model = "rb", "wb", "ne", "msm" or "gformula"} with an empty \code{postc} and 
\code{EMint} is \code{FALSE}, \code{Rcde}, \code{Rpnde}, \code{Rtnde}, \code{Rpnie}, \code{tnie}, 
\code{te} and \code{pm} are estimated.

When \code{model = "msm" or "gformula"} with a non-empty \code{postc}, \code{Rpnde}, \code{Rtnde}, 
\code{Rpnie}, \code{Rtnie},  \code{ERintref}, \code{ERintmed}, \code{ERpnie}, 
\code{ERintref(prop)}, \code{ERintmed(prop)}, \code{ERpnie(prop)}, \code{pm}, \code{int} and 
\code{pe} are replaced by their randomized analogues \code{rRpnde}, \code{rRtnde}, \code{rRpnie}, 
\code{rRtnie}, \code{rERintref}, \code{rERintmed}, \code{rERpnie}, \code{rERintref(prop)}, 
\code{rERintmed(prop)}, \code{rERpnie(prop)}, \code{rpm}, \code{rint} and \code{rpe}.

When \code{model = "iorw"}, \code{Rte}, \code{Rpnde}, \code{Rtnie} and \code{pm} are estimated.
}
\section{Methods (by generic)}{
\itemize{
\item \code{print}: Print the results of \code{cmest} nicely

\item \code{summary}: Summarize the results of \code{cmest} nicely

\item \code{print}: Print the summary of \code{cmest} nicely
}}

\examples{

\dontrun{
library(CMAverse)

# single-mediator case with rb
exp1 <- cmest(data = cma2020, model = "rb", outcome = "contY", 
exposure = "A", mediator = "M2", basec = c("C1", "C2"), 
EMint = TRUE, mreg = list("multinomial"), yreg = "linear", 
astar = 0, a = 1, mval = list("M2_0"), estimation = "paramfunc", 
inference = "delta")
summary(exp1)

# multiple-mediator case with rb
# 10 boots are used for illustration
exp2 <- cmest(data = cma2020, model = "rb", outcome = "contY", 
exposure = "A", mediator = c("M1", "M2"), basec = c("C1", "C2"), 
EMint = TRUE, mreg = list("logistic", "multinomial"), 
yreg = "linear", astar = 0, a = 1, mval = list(0, "M2_0"), 
estimation = "imputation", inference = "bootstrap", nboot = 10,
boot.ci.type = "bca")

# multiple-mediator case with ne
exp3 <- cmest(data = cma2020, model = "ne", outcome = "contY", 
exposure = "A", mediator = c("M1", "M2"), basec = c("C1", "C2"), 
yreg = glm(contY ~ A + M1 + M2 + A*M1 + A*M2 + C1 + C2, family = gaussian, data = cma2020), 
astar = 0, a = 1, mval = list(0, "M2_0"), estimation = "imputation", 
inference = "bootstrap", nboot = 10)

# case control study with msm
exp4 <- cmest(data = cma2020, model = "msm", casecontrol = TRUE, 
yrare = TRUE, outcome = "binY", exposure = "A", 
mediator = c("M1", "M2"), EMint = TRUE, basec = c("C1", "C2"), yreg = "logistic", 
ereg = "logistic", mreg = list(glm(M1 ~ A, family = binomial, 
data = cma2020), nnet::multinom(M2 ~ A, data = cma2020, trace = FALSE)), 
wmnomreg = list(glm(M1 ~ A, family = binomial, data = cma2020), 
nnet::multinom(M2 ~ A + M1, data = cma2020, trace = FALSE)),
wmdenomreg = list(glm(M1 ~ A + C1 + C2, family = binomial, data = cma2020), 
nnet::multinom(M2 ~ A + M1 + C1 + C2, data = cma2020, trace = FALSE)), astar = 0, a = 1, 
mval = list(0, "M2_0"), estimation = "imputation", 
inference = "bootstrap", nboot = 10)
}

}
\references{
Valeri L, VanderWeele TJ (2013). Mediation analysis allowing for
exposure-mediator interactions and causal interpretation: theoretical assumptions and
implementation with SAS and SPSS macros. Psychological Methods. 18(2): 137 - 150.

VanderWeele TJ, Vansteelandt S (2014). Mediation analysis with multiple mediators.
Epidemiologic Methods. 2(1): 95 - 115.

Tchetgen Tchetgen EJ (2013). Inverse odds ratio-weighted estimation for causal
mediation analysis. Statistics in medicine. 32: 4567 - 4580.

Nguyen QC, Osypuk TL, Schmidt NM, Glymour MM, Tchetgen Tchetgen EJ (2015). Practical guidance
for conducting mediation analysis with multiple mediators using inverse odds ratio
weighting. American Journal of Epidemiology. 181(5): 349 - 356.

VanderWeele TJ, Tchetgen Tchetgen EJ (2017). Mediation analysis with time varying
exposures and mediators. Journal of the Royal Statistical Society: Series B (Statistical
Methodology). 79(3): 917 - 938.

Robins JM (1986). A new approach to causal inference in mortality studies with a sustained 
exposure period-Application to control of the healthy worker survivor effect. Mathematical 
Modelling. 7: 1393 - 1512.

Vansteelandt S, Bekaert M, Lange T (2012). Imputation Strategies for the Estimation 
of Natural Direct and Indirect Effects. Epidemiologic Methods. 1(1): 131 - 158.

Steen J, Loeys T, Moerkerke B, Vansteelandt S (2017). Medflex: an R package for
flexible mediation analysis using natural effect models. Journal of Statistical
Software. 76(11).

VanderWeele TJ (2014). A unification of mediation and interaction: a 4-way decomposition. 
Epidemiology. 25(5): 749 - 61.

Imai K, Keele L, Tingley D (2010). A general approach to causal mediation analysis.
Psychological Methods. 15(4): 309 - 334.

Schomaker M, Heumann C (2018). Bootstrap inference when using multiple 
imputation. Statistics in Medicine. 37(14): 2252 - 2266. 

Efron B (1987). Better Bootstrap Confidence Intervals. Journal of the American Statistical 
Association. 82(397): 171-185.
}
\seealso{
\link{ggcmest}, \link{cmdag}, \link{cmsens}, \link{svymultinom}.
}
