% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmsens.R
\name{cmsens}
\alias{cmsens}
\alias{summary.cmsens.me}
\alias{print.summary.cmsens.me}
\alias{plot.cmsens.me}
\title{Sensitivity Analysis For Unmeasured Confounding and Measurement Error}
\usage{
cmsens(
  object = NULL,
  sens = "uc",
  MEmethod = "simex",
  MEvariable = NULL,
  MEvartype = NULL,
  MEerror = NULL,
  lambda = c(0.5, 1, 1.5, 2),
  B = 200,
  nboot.rc = 400
)

\method{summary}{cmsens.me}(cmsens.me)

\method{print}{summary.cmsens.me}(summary.cmsens.me, digits = 4)

\method{plot}{cmsens.me}(cmsens.me)
}
\arguments{
\item{object}{an object of class \code{cmest}.}

\item{sens}{sensitivity analysis for unmeasured confounding or measurement error. 
\code{uc} represents unmeasured confounding and \code{me} represents measurement error.
See \code{Details}.}

\item{MEmethod}{method for measurement error correction. \code{rc} represents regression
calibration and \code{simex} represents SIMEX or MCSIMEX. See \code{Details}.}

\item{MEvariable}{variable measured with error.}

\item{MEvartype}{type of the variable measured with error.
Can be \code{continuous} or \code{categorical} (first 3 letters are enough).}

\item{MEerror}{a vector of standard deviations of the measurement error (when \code{MEvartype}
is \code{continuous}) or a list of misclassification matrices (when \code{MEvartype}
is \code{categorical}).}

\item{lambda}{a vector of lambdas for SIMEX or MCSIMEX. Default is \code{c(0.5, 1, 1.5, 2)}.}

\item{B}{number of simulations for SIMEX or MCSIMEX. Default is \code{200}.}

\item{nboot.rc}{number of boots for variance estimation with regression calibration. Default 
is \code{400}.}
}
\value{
If \code{sens} is \code{uc}, an object of class 'cmest.uc' is returned:
\item{call}{the function call,}
\item{evalues}{the data frame in which the first three columns are the point estimates, 
lower limits of confidence intervals, upper limits of confidence intervals of causal effects on 
the ratio scale and the last three columns are the E-values on the ratio scale for point estimates, 
lower limits of confidence intervals, upper limits of confidence intervals of causal effects,}
If \code{sens} is \code{me}, an object of class 'cmest.me' is returned:
\item{call}{the function call,}
\item{ME}{a list which might contain MEvariable, MEvartype, MEerror, reliability ratio, lambda and B,}
\item{naive}{naive causal mediation analysis results}
\item{sens}{a list of causal mediation analysis results after measurement error correction,}

...
}
\description{
\code{cmsens} is used to conduct sensitivity analysis for unmeasured confounding via 
the \emph{E-value} approach by Vanderweele et al. (2017) and sensitivity analysis for 
measurement error via \emph{regression calibration} by Carroll et al. (1995), 
\emph{SIMEX} by Cook et al. (1994) or \emph{MCSIMEX} by Küchenhoff et al. (2006).
}
\details{
For unmeasured confounding, all E-values are on the (risk or rate) ratio scale. If the causal 
effects are estimated on the difference scale, the point estimates are transformed into ratios.

Sensitivity analysis for measurement error only supports a single variable measured with error.
Regression calibration only supports an independent continuous variable in a regression formula.
SIMEX supports both continuous variables and categorical variables.
}
\section{Methods (by generic)}{
\itemize{
\item \code{summary}: Summarize the results of cmsens.me nicely

\item \code{print}: Print the summary of cmsens.me nicely

\item \code{plot}: Plot the results of cmsens.me with \link[ggplot2]{ggplot}
}}

\examples{
naive <- cmest(data = cma2020, model = "rb", outcome = "Y", 
exposure = "A", mediator = c("M1", "M2"), 
prec = c("C1", "C2"), EMint = TRUE,
mreg = list("logistic", "multinomial"), yreg = "linear",
astar = 0, a = 1, mval = list(0, "M2_0"),
estimation = "imputation", inference = "bootstrap")

exp1 <- cmsens(object = naive, sens = "uc")
exp1

exp2 <- cmsens(object = naive, sens = "me", MEmethod = "rc", 
MEvariable = "C1", MEvartype = "con",
MEerror = c(0.1, 0.2))
summary(exp2)
plot(exp2)

exp3 <- cmsens(object = naive, sens = "me", MEmethod = "simex", 
MEvariable = "M1", MEvartype = "cat",
MEerror = list(matrix(c(0.95,0.05,0.05,0.95), nrow = 2), 
matrix(c(0.9,0.1,0.1,0.9), nrow = 2)))
summary(exp3)

}
\references{
VanderWeele TJ, Ding P. Sensitivity analysis in observational research: introducing the 
E-Value (2017). Annals of Internal Medicine. 167(4): 268 - 274.

Smith LH, VanderWeele TJ. Mediational E-values: Approximate sensitivity analysis for 
unmeasured mediator-outcome confounding (2019). Epidemiology. 30(6): 835 - 837.

Carrol RJ, Ruppert D, Stefanski LA, Crainiceanu C. Measurement Error in Nonlinear Models: 
A Modern Perspective, Second Edition (2006). London: Chapman & Hall.

Cook JR, Stefanski LA. Simulation-extrapolation estimation in parametric measurement error 
models (1994). Journal of the American Statistical Association, 89(428): 1314 - 1328.

Küchenhoff H, Mwalili SM, Lesaffre E. A general method for dealing with misclassification 
in regression: the misclassification SIMEX (2006). Biometrics. 62(1): 85 - 96.

Stefanski LA, Cook JR. Simulation-extrapolation: the measurement error jackknife (1995). 
Journal of the American Statistical Association. 90(432): 1247 - 56.
}
\seealso{
\code{\link{cmdag}}, \code{\link{cmest}}
}
